requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
astor==0.8.1
async-timeout==4.0.3
attrdict==2.0.1
attrs==23.2.0
Babel==2.14.0
backoff==2.2.1
backports.zoneinfo==0.2.1
bce-python-sdk==0.9.5
bcrypt==4.1.2
beautifulsoup4==4.12.3
blinker==1.7.0
build==1.1.1
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.1.1
cssselect==1.2.0
cssutils==2.9.0
cycler==0.12.1
Cython==3.0.9
dataclasses-json==0.6.4
decorator==5.1.1
Deprecated==1.2.14
distro==1.9.0
docx2txt==0.8
easydict==1.13
et-xmlfile==1.1.0
exceptiongroup==1.2.0
faiss-cpu==1.7.1.post2
fastapi==0.110.0
filelock==3.13.1
fire==0.5.0
Flask==3.0.2
flask-babel==4.0.0
flatbuffers==24.3.7
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2024.2.0
future==1.0.0
gast==0.3.3
gitdb==4.0.11
GitPython==3.1.42
google-auth==2.28.2
googleapis-common-protos==1.62.0
graphlib_backport==1.1.0
greenlet==3.0.3
grpcio==1.62.1
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.21.4
humanfriendly==10.0
idna==3.6
imageio==2.34.0
imgaug==0.4.0
importlib-metadata==6.11.0
importlib_resources==6.1.2
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
kubernetes==29.0.0
langchain==0.1.11
langchain-community==0.0.27
langchain-core==0.1.30
langchain-openai==0.0.8
langchain-text-splitters==0.0.1
langsmith==0.1.22
lazy_loader==0.3
lmdb==1.4.1
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.7.5
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.1
numpy==1.24.4
oauthlib==3.2.2
onnxruntime==1.17.1
openai==1.13.3
openpyxl==3.1.2
opentelemetry-api==1.23.0
opentelemetry-exporter-otlp-proto-common==1.23.0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
opentelemetry-instrumentation==0.44b0
opentelemetry-instrumentation-asgi==0.44b0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-proto==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-semantic-conventions==0.44b0
opentelemetry-util-http==0.44b0
opt-einsum==3.3.0
orjson==3.9.15
overrides==7.7.0
packaging==23.2
pandas==2.0.3
pdf2docx==0.5.8
pillow==10.2.0
pkgutil_resolve_name==1.3.10
posthog==3.5.0
premailer==3.10.0
prettytable==3.10.0
protobuf==3.20.2
psutil==5.9.8
pulsar-client==3.4.0
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyclipper==1.3.0.post5
pycryptodome==3.20.0
pydantic==2.6.3
pydantic_core==2.16.3
pydeck==0.8.1b0
Pygments==2.17.2
PyMuPDF==1.20.2
pyparsing==3.1.2
PyPika==0.48.9
pyproject_hooks==1.0.0
pyreadline3==3.4.1
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.0.1
pytz==2024.1
PyWavelets==1.4.1
PyYAML==6.0.1
rapidfuzz==3.6.2
rarfile==4.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
shapely==2.0.3
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.28
starlette==0.36.3
streamlit==1.31.1
streamlit-feedback==0.1.3
sympy==1.12
tenacity==8.2.3
termcolor==2.4.0
threadpoolctl==3.3.0
tifffile==2023.7.10
tiktoken==0.6.0
tokenizers==0.15.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
ujson==5.9.0
urllib3==2.2.1
uvicorn==0.28.0
validators==0.22.0
visualdl==2.5.3
watchdog==4.0.0
watchfiles==0.21.0
wcwidth==0.2.13
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.0.1
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
